backend/requirements.txt
Django==4.2.8
django-cors-headers==4.3.1
werkzeug==2.0.1
gunicorn==21.2.0
boto3==1.34.6
python-dotenv==1.0.0